+++ modules/automation-account/locals.tf
@@ -1,6 +1,2 @@
 locals {
-  default_tags = {
-    env   = var.environment
-    stack = var.stack
-  }
 }