+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased 
+
+Fixed 
+  * AZ-674 : fix this bug  https://claranet.slack.com/archives/CGE40D3KL/p1643202262270200
+
 # v5.0.0 - 2022-01-18
 
 Breaking