+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Breaking
   * AZ-515: Option to use Azure CAF naming provider to name resources
 
+Added
+  * AZ-615: Add an option to enable or disable default tags
+
 Changed
   * AZ-589: Bump `diagnostics` module to latest `v5.0.0`