+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# unreleased
+# v5.0.0 - 2022-03-11
 
 Breaking
   * AZ-515: Option to use Azure CAF naming provider to name resources